% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleApi__HttpBody}
\alias{GoogleApi__HttpBody}
\title{Google Cloud Machine Learning Engine Objects 
An API to enable creating and using machine learning models.}
\usage{
GoogleApi__HttpBody(data = NULL, contentType = NULL)
}
\arguments{
\item{data}{HTTP body binary data}

\item{contentType}{The HTTP Content-Type string representing the content type of the body}
}
\value{
GoogleApi__HttpBody object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:02:30
filename: /Users/mark/dev/R/autoGoogleAPI/googlemlv1beta1.auto/R/ml_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
GoogleApi__HttpBody Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Message that represents an arbitrary HTTP body. It should only be used forpayload formats that can't be represented as JSON, such as raw binary oran HTML page.This message can be used both in streaming and non-streaming API methods inthe request as well as the response.It can be used as a top-level request field, which is convenient if onewants to extract parameters from either the URL or HTTP template into therequest fields and also want access to the raw HTTP body.Example:    message GetResourceRequest {      // A unique request id.      string request_id = 1;      // The raw HTTP body is bound to this field.      google.api.HttpBody http_body = 2;    }    service ResourceService {      rpc GetResource(GetResourceRequest) returns (google.api.HttpBody);      rpc UpdateResource(google.api.HttpBody) returns (google.protobuf.Empty);    }Example with streaming methods:    service CaldavService {      rpc GetCalendar(stream google.api.HttpBody)        returns (stream google.api.HttpBody);      rpc UpdateCalendar(stream google.api.HttpBody)        returns (stream google.api.HttpBody);    }Use of this type only changes how the request and response bodies arehandled, all other features will continue to work unchanged.
}
