% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__ManualScaling}
\alias{GoogleCloudMlV1__ManualScaling}
\title{GoogleCloudMlV1__ManualScaling Object}
\usage{
GoogleCloudMlV1__ManualScaling(nodes = NULL)
}
\arguments{
\item{nodes}{The number of nodes to allocate for this model}
}
\value{
GoogleCloudMlV1__ManualScaling object
}
\description{
GoogleCloudMlV1__ManualScaling Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Options for manually scaling a model.
}
