% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__OperationMetadata}
\alias{GoogleCloudMlV1__OperationMetadata}
\title{GoogleCloudMlV1__OperationMetadata Object}
\usage{
GoogleCloudMlV1__OperationMetadata(createTime = NULL, modelName = NULL,
  version = NULL, endTime = NULL, operationType = NULL,
  startTime = NULL, isCancellationRequested = NULL)
}
\arguments{
\item{createTime}{The time the operation was submitted}

\item{modelName}{Contains the name of the model associated with the operation}

\item{version}{Contains the version associated with the operation}

\item{endTime}{The time operation processing completed}

\item{operationType}{The operation type}

\item{startTime}{The time operation processing started}

\item{isCancellationRequested}{Indicates whether a request to cancel this operation has been made}
}
\value{
GoogleCloudMlV1__OperationMetadata object
}
\description{
GoogleCloudMlV1__OperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the metadata of the long-running operation.
}
