% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1beta1__CancelJobRequest}
\alias{GoogleCloudMlV1beta1__CancelJobRequest}
\title{GoogleCloudMlV1beta1__CancelJobRequest Object}
\usage{
GoogleCloudMlV1beta1__CancelJobRequest()
}
\value{
GoogleCloudMlV1beta1__CancelJobRequest object
}
\description{
GoogleCloudMlV1beta1__CancelJobRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for the CancelJob method.
}
