% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1beta1__GetConfigResponse}
\alias{GoogleCloudMlV1beta1__GetConfigResponse}
\title{GoogleCloudMlV1beta1__GetConfigResponse Object}
\usage{
GoogleCloudMlV1beta1__GetConfigResponse(serviceAccountProject = NULL,
  serviceAccount = NULL)
}
\arguments{
\item{serviceAccountProject}{The project number for `service_account`}

\item{serviceAccount}{The service account Cloud ML uses to access resources in the project}
}
\value{
GoogleCloudMlV1beta1__GetConfigResponse object
}
\description{
GoogleCloudMlV1beta1__GetConfigResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Returns service account information associated with a project.
}
