% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1beta1__HyperparameterSpec}
\alias{GoogleCloudMlV1beta1__HyperparameterSpec}
\title{GoogleCloudMlV1beta1__HyperparameterSpec Object}
\usage{
GoogleCloudMlV1beta1__HyperparameterSpec(maxParallelTrials = NULL,
  goal = NULL, hyperparameterMetricTag = NULL, params = NULL,
  maxTrials = NULL)
}
\arguments{
\item{maxParallelTrials}{Optional}

\item{goal}{Required}

\item{hyperparameterMetricTag}{Optional}

\item{params}{Required}

\item{maxTrials}{Optional}
}
\value{
GoogleCloudMlV1beta1__HyperparameterSpec object
}
\description{
GoogleCloudMlV1beta1__HyperparameterSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a set of hyperparameters to optimize.
}
