% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1beta1__Job}
\alias{GoogleCloudMlV1beta1__Job}
\title{GoogleCloudMlV1beta1__Job Object}
\usage{
GoogleCloudMlV1beta1__Job(endTime = NULL, startTime = NULL,
  predictionOutput = NULL, trainingOutput = NULL, trainingInput = NULL,
  createTime = NULL, state = NULL, predictionInput = NULL,
  errorMessage = NULL, jobId = NULL)
}
\arguments{
\item{endTime}{Output only}

\item{startTime}{Output only}

\item{predictionOutput}{The current prediction job result}

\item{trainingOutput}{The current training job result}

\item{trainingInput}{Input parameters to create a training job}

\item{createTime}{Output only}

\item{state}{Output only}

\item{predictionInput}{Input parameters to create a prediction job}

\item{errorMessage}{Output only}

\item{jobId}{Required}
}
\value{
GoogleCloudMlV1beta1__Job object
}
\description{
GoogleCloudMlV1beta1__Job Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a training or prediction job.
}
