% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1beta1__ListJobsResponse}
\alias{GoogleCloudMlV1beta1__ListJobsResponse}
\title{GoogleCloudMlV1beta1__ListJobsResponse Object}
\usage{
GoogleCloudMlV1beta1__ListJobsResponse(nextPageToken = NULL, jobs = NULL)
}
\arguments{
\item{nextPageToken}{Optional}

\item{jobs}{The list of jobs}
}
\value{
GoogleCloudMlV1beta1__ListJobsResponse object
}
\description{
GoogleCloudMlV1beta1__ListJobsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for the ListJobs method.
}
