% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1beta1__ListModelsResponse}
\alias{GoogleCloudMlV1beta1__ListModelsResponse}
\title{GoogleCloudMlV1beta1__ListModelsResponse Object}
\usage{
GoogleCloudMlV1beta1__ListModelsResponse(nextPageToken = NULL,
  models = NULL)
}
\arguments{
\item{nextPageToken}{Optional}

\item{models}{The list of models}
}
\value{
GoogleCloudMlV1beta1__ListModelsResponse object
}
\description{
GoogleCloudMlV1beta1__ListModelsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for the ListModels method.
}
