% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1beta1__ListVersionsResponse}
\alias{GoogleCloudMlV1beta1__ListVersionsResponse}
\title{GoogleCloudMlV1beta1__ListVersionsResponse Object}
\usage{
GoogleCloudMlV1beta1__ListVersionsResponse(versions = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{versions}{The list of versions}

\item{nextPageToken}{Optional}
}
\value{
GoogleCloudMlV1beta1__ListVersionsResponse object
}
\description{
GoogleCloudMlV1beta1__ListVersionsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for the ListVersions method.
}
