% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1beta1__Model}
\alias{GoogleCloudMlV1beta1__Model}
\title{GoogleCloudMlV1beta1__Model Object}
\usage{
GoogleCloudMlV1beta1__Model(description = NULL,
  onlinePredictionLogging = NULL, defaultVersion = NULL, regions = NULL,
  name = NULL)
}
\arguments{
\item{description}{Optional}

\item{onlinePredictionLogging}{Optional}

\item{defaultVersion}{Output only}

\item{regions}{Optional}

\item{name}{Required}
}
\value{
GoogleCloudMlV1beta1__Model object
}
\description{
GoogleCloudMlV1beta1__Model Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a machine learning solution.A model can have multiple versions, each of which is a deployed, trainedmodel ready to receive prediction requests. The model itself is just acontainer.
}
