% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1beta1__ParameterSpec}
\alias{GoogleCloudMlV1beta1__ParameterSpec}
\title{GoogleCloudMlV1beta1__ParameterSpec Object}
\usage{
GoogleCloudMlV1beta1__ParameterSpec(minValue = NULL, discreteValues = NULL,
  scaleType = NULL, maxValue = NULL, type = NULL,
  categoricalValues = NULL, parameterName = NULL)
}
\arguments{
\item{minValue}{Required if type is `DOUBLE` or `INTEGER`}

\item{discreteValues}{Required if type is `DISCRETE`}

\item{scaleType}{Optional}

\item{maxValue}{Required if typeis `DOUBLE` or `INTEGER`}

\item{type}{Required}

\item{categoricalValues}{Required if type is `CATEGORICAL`}

\item{parameterName}{Required}
}
\value{
GoogleCloudMlV1beta1__ParameterSpec object
}
\description{
GoogleCloudMlV1beta1__ParameterSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a single hyperparameter to optimize.
}
