% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1beta1__PredictionInput}
\alias{GoogleCloudMlV1beta1__PredictionInput}
\title{GoogleCloudMlV1beta1__PredictionInput Object}
\usage{
GoogleCloudMlV1beta1__PredictionInput(region = NULL, versionName = NULL,
  modelName = NULL, outputPath = NULL, maxWorkerCount = NULL,
  uri = NULL, dataFormat = NULL, runtimeVersion = NULL,
  inputPaths = NULL)
}
\arguments{
\item{region}{Required}

\item{versionName}{Use this field if you want to specify a version of the model to use}

\item{modelName}{Use this field if you want to use the default version for the specified}

\item{outputPath}{Required}

\item{maxWorkerCount}{Optional}

\item{uri}{Use this field if you want to specify a Google Cloud Storage path for}

\item{dataFormat}{Required}

\item{runtimeVersion}{Optional}

\item{inputPaths}{Required}
}
\value{
GoogleCloudMlV1beta1__PredictionInput object
}
\description{
GoogleCloudMlV1beta1__PredictionInput Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents input parameters for a prediction job.
}
