% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1beta1__SetDefaultVersionRequest}
\alias{GoogleCloudMlV1beta1__SetDefaultVersionRequest}
\title{GoogleCloudMlV1beta1__SetDefaultVersionRequest Object}
\usage{
GoogleCloudMlV1beta1__SetDefaultVersionRequest()
}
\value{
GoogleCloudMlV1beta1__SetDefaultVersionRequest object
}
\description{
GoogleCloudMlV1beta1__SetDefaultVersionRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for the SetDefaultVersion request.
}
