% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1beta1__TrainingInput}
\alias{GoogleCloudMlV1beta1__TrainingInput}
\title{GoogleCloudMlV1beta1__TrainingInput Object}
\usage{
GoogleCloudMlV1beta1__TrainingInput(runtimeVersion = NULL,
  pythonModule = NULL, workerType = NULL, args = NULL, region = NULL,
  parameterServerType = NULL, scaleTier = NULL, jobDir = NULL,
  hyperparameters = NULL, parameterServerCount = NULL, packageUris = NULL,
  workerCount = NULL, masterType = NULL)
}
\arguments{
\item{runtimeVersion}{Optional}

\item{pythonModule}{Required}

\item{workerType}{Optional}

\item{args}{Optional}

\item{region}{Required}

\item{parameterServerType}{Optional}

\item{scaleTier}{Required}

\item{jobDir}{Optional}

\item{hyperparameters}{Optional}

\item{parameterServerCount}{Optional}

\item{packageUris}{Required}

\item{workerCount}{Optional}

\item{masterType}{Optional}
}
\value{
GoogleCloudMlV1beta1__TrainingInput object
}
\description{
GoogleCloudMlV1beta1__TrainingInput Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents input parameters for a training job.
}
