% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1beta1__TrainingOutput}
\alias{GoogleCloudMlV1beta1__TrainingOutput}
\title{GoogleCloudMlV1beta1__TrainingOutput Object}
\usage{
GoogleCloudMlV1beta1__TrainingOutput(completedTrialCount = NULL,
  isHyperparameterTuningJob = NULL, consumedMLUnits = NULL, trials = NULL)
}
\arguments{
\item{completedTrialCount}{The number of hyperparameter tuning trials that completed successfully}

\item{isHyperparameterTuningJob}{Whether this job is a hyperparameter tuning job}

\item{consumedMLUnits}{The amount of ML units consumed by the job}

\item{trials}{Results for individual Hyperparameter trials}
}
\value{
GoogleCloudMlV1beta1__TrainingOutput object
}
\description{
GoogleCloudMlV1beta1__TrainingOutput Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents results of a training job. Output only.
}
