% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1beta1__Version}
\alias{GoogleCloudMlV1beta1__Version}
\title{GoogleCloudMlV1beta1__Version Object}
\usage{
GoogleCloudMlV1beta1__Version(runtimeVersion = NULL, lastUseTime = NULL,
  description = NULL, deploymentUri = NULL, isDefault = NULL,
  createTime = NULL, manualScaling = NULL, name = NULL)
}
\arguments{
\item{runtimeVersion}{Optional}

\item{lastUseTime}{Output only}

\item{description}{Optional}

\item{deploymentUri}{Required}

\item{isDefault}{Output only}

\item{createTime}{Output only}

\item{manualScaling}{Optional}

\item{name}{Required}
}
\value{
GoogleCloudMlV1beta1__Version object
}
\description{
GoogleCloudMlV1beta1__Version Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a version of the model.Each version is a trained model deployed in the cloud, ready to handleprediction requests. A model can have multiple versions. You can getinformation about all of the versions of a given model by calling[projects.models.versions.list](/ml/reference/rest/v1beta1/projects.models.versions/list).
}
