% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1beta1_HyperparameterOutput_HyperparameterMetric}
\alias{GoogleCloudMlV1beta1_HyperparameterOutput_HyperparameterMetric}
\title{GoogleCloudMlV1beta1_HyperparameterOutput_HyperparameterMetric Object}
\usage{

  GoogleCloudMlV1beta1_HyperparameterOutput_HyperparameterMetric(trainingStep = NULL,
  objectiveValue = NULL)
}
\arguments{
\item{trainingStep}{The global training step for this metric}

\item{objectiveValue}{The objective value at this training step}
}
\value{
GoogleCloudMlV1beta1_HyperparameterOutput_HyperparameterMetric object
}
\description{
GoogleCloudMlV1beta1_HyperparameterOutput_HyperparameterMetric Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An observed value of a metric.
}
