% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleLongrunning__ListOperationsResponse}
\alias{GoogleLongrunning__ListOperationsResponse}
\title{GoogleLongrunning__ListOperationsResponse Object}
\usage{
GoogleLongrunning__ListOperationsResponse(operations = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{operations}{A list of operations that matches the specified filter in the request}

\item{nextPageToken}{The standard List next-page token}
}
\value{
GoogleLongrunning__ListOperationsResponse object
}
\description{
GoogleLongrunning__ListOperationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response message for Operations.ListOperations.
}
