% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleLongrunning__Operation.metadata}
\alias{GoogleLongrunning__Operation.metadata}
\title{GoogleLongrunning__Operation.metadata Object}
\usage{
GoogleLongrunning__Operation.metadata()
}
\value{
GoogleLongrunning__Operation.metadata object
}
\description{
GoogleLongrunning__Operation.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Service-specific metadata associated with the operation.  It typicallycontains progress information and common metadata such as create time.Some services might not provide such metadata.  Any method that returns along-running operation should document the metadata type, if any.
}
\seealso{
Other GoogleLongrunning__Operation functions: \code{\link{GoogleLongrunning__Operation.response}},
  \code{\link{GoogleLongrunning__Operation}}
}
