% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleLongrunning__Operation.response}
\alias{GoogleLongrunning__Operation.response}
\title{GoogleLongrunning__Operation.response Object}
\usage{
GoogleLongrunning__Operation.response()
}
\value{
GoogleLongrunning__Operation.response object
}
\description{
GoogleLongrunning__Operation.response Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The normal response of the operation in case of success.  If the originalmethod returns no data on success, such as `Delete`, the response is`google.protobuf.Empty`.  If the original method is standard`Get`/`Create`/`Update`, the response should be the resource.  For othermethods, the response should have the type `XxxResponse`, where `Xxx`is the original method name.  For example, if the original method nameis `TakeSnapshot()`, the inferred response type is`TakeSnapshotResponse`.
}
\seealso{
Other GoogleLongrunning__Operation functions: \code{\link{GoogleLongrunning__Operation.metadata}},
  \code{\link{GoogleLongrunning__Operation}}
}
