% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleProtobuf__Empty}
\alias{GoogleProtobuf__Empty}
\title{GoogleProtobuf__Empty Object}
\usage{
GoogleProtobuf__Empty()
}
\value{
GoogleProtobuf__Empty object
}
\description{
GoogleProtobuf__Empty Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A generic empty message that you can re-use to avoid defining duplicatedempty messages in your APIs. A typical example is to use it as the requestor the response type of an API method. For instance:    service Foo {      rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty);    }The JSON representation for `Empty` is empty JSON object `{}`.
}
