% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\docType{package}
\name{ml_googleAuthR}
\alias{ml_googleAuthR}
\alias{ml_googleAuthR-package}
\title{Google Cloud Machine Learning Engine
An API to enable creating and using machine learning models.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:02:30
filename: /Users/mark/dev/R/autoGoogleAPI/googlemlv1beta1.auto/R/ml_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}
}
