% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{projects.getConfig}
\alias{projects.getConfig}
\title{Get the service account information associated with your project. You needthis information in order to grant the service account persmissions forthe Google Cloud Storage location where you put your model training codefor training the model with Google Cloud Machine Learning.}
\usage{
projects.getConfig(name)
}
\arguments{
\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/ml/}{Google Documentation}
}
