% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{projects.predict}
\alias{projects.predict}
\title{Performs prediction on the data in the request.**** REMOVE FROM GENERATED DOCUMENTATION}
\usage{
projects.predict(GoogleCloudMlV1beta1__PredictRequest, name)
}
\arguments{
\item{GoogleCloudMlV1beta1__PredictRequest}{The \link{GoogleCloudMlV1beta1__PredictRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/ml/}{Google Documentation}

Other GoogleCloudMlV1beta1__PredictRequest functions: \code{\link{GoogleCloudMlV1beta1__PredictRequest}}
}
